Moved price feed functions from OracleAPI to PriceFeedAPI. Moved medianizer query from SystemAPI to OracleAPI.
