+++ test/integration/api/SystemAPI.spec.ts
@@ -30,7 +30,6 @@ import { Address, Web3Utils } from 'set-protocol-utils';
 import * as setProtocolUtils from 'set-protocol-utils';
 import {
   CoreContract,
-  MedianContract,
   RebalanceAuctionModuleContract,
   RebalancingSetTokenFactoryContract,
   SetTokenFactoryContract,
@@ -52,13 +51,10 @@ import {
   SystemTimeLockPeriodState,
 } from '@src/types/common';
 import {
-  addPriceFeedOwnerToMedianizer,
   addPriceLibraryAsync,
   deployBaseContracts,
-  deployMedianizerAsync,
   deployWhitelistContract,
   registerExchange,
-  updateMedianizerPriceAsync,
 } from '@test/helpers';
 
 ChaiSetup.configure();
@@ -406,40 +402,6 @@ describe('SystemAPI', () => {
         DEFAULT_ACCOUNT,
       ];
       expect(JSON.stringify(priceLibraries)).to.equal(JSON.stringify(expectedPriceLibraries));
-    });
-  });
-
-  describe('getFeedPriceAsync', async () => {
-    let subjectMedianizerAddress: Address;
-
-    let btcMedianizer: MedianContract;
-    let btcPrice: BigNumber;
-
-    beforeEach(async () => {
-      btcMedianizer = await deployMedianizerAsync(web3);
-      await addPriceFeedOwnerToMedianizer(btcMedianizer, DEFAULT_ACCOUNT);
-
-      btcPrice = new BigNumber(4082 * 10 ** 18);
-      await updateMedianizerPriceAsync(
-        web3,
-        btcMedianizer,
-        btcPrice,
-        SetTestUtils.generateTimestamp(1000),
-      );
-
-      subjectMedianizerAddress = btcMedianizer.address;
-    });
-
-    async function subject(): Promise<BigNumber> {
-      return await systemAPI.getFeedPriceAsync(
-        subjectMedianizerAddress
-      );
-    }
-
-    test('gets the correct price', async () => {
-      const price = await subject();
-
-      expect(price).to.bignumber.equal(btcPrice);
     });
   });
 });
